+++ CHANGELOG.md
@@ -3,12 +3,13 @@
 ## Unreleased
 
 ### Added
+- Can now select which subfields will use autocomplete.
 - Addresses now include an optional [`name`](https://plugins.doublesecretagency.com/mapbox/models/address-model/#name) subfield. ([#4](https://github.com/doublesecretagency/craft-mapbox/issues/4))
 - Addresses now include an optional [`neighborhood`](https://plugins.doublesecretagency.com/mapbox/models/address-model/#neighborhood) subfield.
 - Addresses now include an optional [`county`](https://plugins.doublesecretagency.com/mapbox/models/address-model/#county) subfield.
 - Addresses now include an optional [`mapboxId`](https://plugins.doublesecretagency.com/mapbox/models/address-model/#mapboxid) subfield.
-- Can now select which subfields will use autocomplete.
-- Added [`fieldParams`](https://plugins.doublesecretagency.com/mapbox/getting-started/config/#fieldparams) setting to configure the language of Address fields. ([#7](https://github.com/doublesecretagency/craft-mapbox/pull/7))
+- Address fields will now auto-detect the language of the currently logged-in user.
+- Address fields' language can be manually overridden with the new [`fieldParams`](https://plugins.doublesecretagency.com/mapbox/getting-started/config/#fieldparams) setting. ([#7](https://github.com/doublesecretagency/craft-mapbox/pull/7))
 
 ### Changed
 - Replaced the underlying search API used in the Address field.